First Naive implementation of GSPAT supporting more than 1024 transducers.
Very Naive extension to GS-PAT to support more boards.

The current implementation only works for multiples of 256 (2,4,8,16 boards). The implementation is also hardcoded. See constant definition in "hologramSolver_v2.cl" and "HologramSolverCL.cpp":

#define NUM_TRANSDUCERS_PER_GROUP 1

Please note both definitions must be changed and match, to adapt to higher number of transducers (1 allows for up to 1024 transducers, 2 allows for 2048, 4 allows for 4096...)

THIS IMPLEMENTATION SHOULD ONLY BE USED AS  A BASELINE TO IMPROVE AGAINST.