Updated common Solver interface (GSPAT_Solver.h), to add custom configuration parameters to each solver.
GSPAT_Solver.h now include a method setConfigParameters, which allows the client to set an arbitrary number of parameters to the specific solver used. This is necessary for advanced solvers allowing for arbitrary transducer normals, integrating reflectors (BEM), and whatever else comes to mind in the future.
The base class provides a default implementation (doing nothing), so old solvers should not be affected (but will need to be recompiled)